task1/requirements.txt
torch
ultralytics
numpy
opencv-python
Pillow
mtcnn
tqdm
argparse